requirements.txt
fastapi
uvicorn
requests
PyYAML
docling
pypdfium2